3.2.8b3 minor fix to remove warning on production compile
git-svn-id: http://squeakvm.org/svn/squeak/trunk@454 fa1542d4-bde8-0310-ad64-8ed1123d492a

+++ platforms/Mac OS/vm/sqMacImageIO.c
@@ -6,7 +6,7 @@
 *   AUTHOR:  John Maloney, John McIntosh, and others.
 *   ADDRESS: 
 *   EMAIL:   johnmci@smalltalkconsulting.com
-*   RCSID:   $Id: sqMacImageIO.c,v 1.2 2002/02/23 11:25:39 johnmci Exp $
+*   RCSID:   $Id: sqMacImageIO.c,v 1.3 2002/08/06 22:35:27 johnmci Exp $
 *
 *   NOTES: 
 *  Feb 22nd, 2002, JMM moved code into 10 other files, see sqMacMain.c for comments
@@ -118,7 +118,7 @@ sqImageFile sqImageFileOpen(char *fileName, char *mode) {
     char cFileName[1024];
     sqImageFile remember;
     
-    sqFilenameFromStringOpen(cFileName, fileName, strlen(fileName));
+    sqFilenameFromStringOpen(cFileName,(long) fileName, strlen(fileName));
     remember = fopen(cFileName, mode);
     if (remember == null) 
         return null;
